Fix race issue in gridview resetAllSlides()
There was a race issue in this call which occurred when loadBatch()
was executing at the same time that resetAllSlides() was called,
causing incorrect slide order insertion. This has been fixed by
waiting for the current loadBatch() to complete before resetting.

It might be better to abort loadBatch() instead of waiting for it?