[refactor] Simplify batch loading (#101)
* remove repeated code from grid-view loadBatch()

* reenable graying-out navigation buttons at end of slide deck

* add bug fixes and console cleanup

- Fixed loss of double-tap handler in grid view.
- Removed unneeded console.log() debug statements.

* Fix race issue in gridview resetAllSlides()

There was a race issue in this call which occurred when loadBatch()
was executing at the same time that resetAllSlides() was called,
causing incorrect slide order insertion. This has been fixed by
waiting for the current loadBatch() to complete before resetting.

It might be better to abort loadBatch() instead of waiting for it?

---------

Co-authored-by: Lincoln Stein <lstein@gmail.com>